+++ src/assertion/purchaseAssertion.js
