+++ src/schemas/product.schema.json
@@ -1,0 +1,52 @@
+{
+  "$ref": "#/definitions/Product",
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "definitions": {
+    "ProductCategory": {
+      "enum": [
+        "Food",
+        "Beverage",
+        "Clothes",
+        "Furniture",
+        "Tools"
+      ],
+      "type": "string"
+    },
+    "Product": {
+      "type": "object",
+      "additionalProperties": false,
+      "properties": {
+        "name": {
+          "type": "string",
+          "minLength": 4,
+          "maxLength": 32
+        },
+        "category": {
+          "$ref": "#/definitions/Types.ContactType"
+        },
+        "qty": {
+          "type": "number",
+          "minimum": 1
+        },
+        "price": {
+          "type": "number",
+          "minimum": 100
+        },
+        "sku": {
+          "type": "string"
+        },
+        "fileId": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "name",
+        "category",
+        "qty",
+        "price",
+        "sku",
+        "fileId"
+      ]
+    }
+  }
+}
